Update map_gen to use 28-bit (not 16-bit) seeds by default
map_gen now uses 28-bit seeds by default, so that we do not have
to worry about "birthday paradox" collisions if we make a map every
day for a year.

Also, update readme.md to point out this only compiles in an old
version of mingw + msys I have in a VMware virtual machine.