HOTFIX: Make code run after January 19, 2038
The underlying code which map_gen uses will run fine any time in
the future -- as long as time_t is 64-bit.  Since the system we
use to compile it on has a 32-bit time_t, the entire program
would had stopped running on January 19, 2038 (January 18 in many
parts of the world).

I have fixed the code to continue running after that date, albeit
with the code complaning we are before 1980, and with timestamps
using an incorrect negative number instead of being human-readable.

But, this code no longer has a fatal Y2038 problem.