Update for 2021-05-20
In this update, we allow seeds to be specified with the '-s' argument.

For example: map_gen.exe -s 12345

Seeds specified this way *must* contain *only* numbers or lower case
letters ("-" is also allowed in a seed, but can *not* be the first
character of the seed).

Seeds specified this way are in a different namespace than seeds specified
with the -number or with the help of the file "seed2.txt".  For example,
"map_gen.exe -1" generates a different map than "map_gen.exe -s 1"

The seed string specified with "-s" can be up to 230 characters long.